backend/requirements.txt
fastapi==0.115.4
google-generativeai==0.8.3
pillow==11.0.0
requests==2.25.1
python-multipart==0.0.17
python-dotenv==1.0.1